+++ src/game/game_loop.rs
@@ -43,7 +43,7 @@ fn has_snake_reached_apple(snake: &Snake, apple: &Apple) -> bool {
 
 pub async fn run_game_loop(grid_width: u32, grid_height: u32, cell_size: f32) {
     let mut snake = Snake::new(Coordinates::new(1, 1), Orientation::East);
-    let mut apple = Apple::random(grid_width as i32, grid_height as i32, &snake);
+    let mut apple = Apple::random(grid_width, grid_height, &snake);
 
     let mut state = GameState::Running;
     let mut timer = 0.0;
@@ -79,7 +79,7 @@ pub async fn run_game_loop(grid_width: u32, grid_height: u32, cell_size: f32) {
                         state = GameState::GameOver;
                     } else if has_snake_reached_apple(&snake, &apple) {
                         grow = true;
-                        apple = Apple::random(grid_width as i32, grid_height as i32, &snake);
+                        apple = Apple::random(grid_width, grid_height, &snake);
                     }
 
                     snake.advance(grow);
@@ -89,7 +89,7 @@ pub async fn run_game_loop(grid_width: u32, grid_height: u32, cell_size: f32) {
                 // press Enter to restart
                 if is_key_pressed(KeyCode::Enter) {
                     snake = Snake::new(Coordinates::new(1, 1), Orientation::East);
-                    apple = Apple::random(grid_width as i32, grid_height as i32, &snake);
+                    apple = Apple::random(grid_width, grid_height, &snake);
                     state = GameState::Running;
                     timer = 0.0;
                 }
